Protect against deadlocks even when code behaves badly.
In the case that a generator starts a goroutine that both try to get the cycle lock, there is a slim chance that both can get it and that could potentially lead to a deadlock.

In reality, generators shouldn't be starting goroutines so this shouldn't have been a problem, but I want this to have the highest standards.