Introduce the concept of loose dependencies
The default behavior allows only a single dependency, be it a concrete value or a generator, for a slot. This is safer for production code.

Loose dependencies is the opposite of strict. The intent is that this allows for mainly test code to override a default set of dependencies to exercise use cases easier without having to create bespoke dependency initializers.