Security update
Malicious code could be used to have distant effects by filling dependencies from a child context that could, in theory, change the effect of the generators.

This locks the generators to use the context they were originally constructed in and prohibits the access of context values that may be dangerous.

This is done in a way that respects timeouts and errors that are on the passed-in context.