Add TTL check and new test in cache package
The commit introduces a check that prevents setting of time-to-live (TTL) if the value is less than or equal to zero in cache package. It enhances cache management by skipping unnecessary data store operations for non-persistent objects. Additionally, a new test function `Test_Cache_ObjectTTL_Zero` has been added to ensure correctness of this behavior.