Introduce the concept of pre-refreshing the cache + some overall simplifications
Prior to this, cache entries would exist until they expire at which point they would be refreshed. The potential issue with this is the call to the generator function can cause processing to be blocked. While this is no different than a simple cache miss, this can be improved.

The new feature introduced is the concept of a pre-refresh. Once a cache entry goes beyond `RefreshPercentage` of its age, it will be refreshed in the background. While the refresh is happening, the existing, still valid, cached results are returned. This ensures that for high-volume cache items, there won't be a case where the result isn't already in the cache and valid. This will speed up overall program execution.

Additionally, this removes the ill-thought-out cache locking. It doesn't offer the level of nuance required for a true distributed cache. There is a new package, go-rediscache, that offers similar functionality, but optimized for distributed caching.